google-drive/requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
docutils==0.21.2
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
Kivy==2.3.0
Kivy-Garden==0.1.5
kivymd==1.2.0
oauthlib==3.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.2
